+++ src/internet.jl
@@ -1,71 +1,32 @@
 
-uri_pages = ["index", "home", "search", "main", "post", "homepage", "category", "register", "login", "faq", "about", "terms","privacy", "author"]
-uri_paths = ["app", "main", "wp-content", "search", "category", "tag", "categories", "tags", "blog", "posts", "list", "explore"]
-uri_extensions = [".html", ".html", ".html", ".htm", ".htm", ".php", ".php", ".jsp", ".asp"]
+const uri_pages = ["index", "home", "search", "main", "post", "homepage", "category", "register", "login", "faq", "about", "terms","privacy", "author"]
+const uri_paths = ["app", "main", "wp-content", "search", "category", "tag", "categories", "tags", "blog", "posts", "list", "explore"]
+const uri_extensions = [".html", ".html", ".html", ".htm", ".htm", ".php", ".php", ".jsp", ".asp"]
 
+pixels()::Int = rand(1:1024)
+user_name()::String = bothify(executor(data["faker"]["internet"]["user_name_formats"]))
+domain_suffix()::String = executor(data["faker"]["internet"]["domain_suffix"])
+safe_email(items::Tuple=(user_name(),"@example.", domain_suffix()))::String = reduce(string, items)
+free_email_domain()::String = executor(data["faker"]["internet"]["free_email"])
+free_email(items::Tuple=(user_name(),"@",free_email_domain()))::String = reduce(string, items)
+domain_word()::String = lowercase(join(map( x -> length(x)>0 ? x[1] : "" , split(Faker.company(), [',', ' ','-'])),""))
+domain_name(items::Tuple=(domain_word(),".",domain_suffix()))::String = reduce(string, items)
+company_email(items::Tuple=(user_name(),"@",domain_name()))::String = reduce(string, items)
 
-pixels()= rand(1:1024)
+email()::String = executor(data["faker"]["internet"]["email_formats"])
+url()::String = executor(data["faker"]["internet"]["url_formats"])
 
-user_name()= bothify(executor(data["faker"]["internet"]["user_name_formats"]))
+ipv4()::String = string(rand( 172:255),".",rand( 0:255),".",rand(0:255),".",rand(0:255))
+ipv6()::String = join(map( x -> string(rand(4100:65535),base=16) ,1:8), ":")
+mac_address()::String = join(map( x -> string(Base.hex(unsigned(rand( 0x00:0xff)),2,false)) ,1:6), ":")
+uri_page()::String = rand(uri_pages)
 
-domain_suffix()= executor(data["faker"]["internet"]["domain_suffix"])
-safe_email()= "$(user_name())@example.$(domain_suffix())"
-
-free_email_domain() = executor(data["faker"]["internet"]["free_email"])
-free_email()= "$(user_name())@$(free_email_domain())"
-
-
-
-domain_word()=
-   ( company_elements = split(company(), [',', ' ','-']);
-    sal = "";
-    for i in company_elements
-               if length(i)>0
-                   sal*="$(i[1])"
-               end
-          end;
-
-    lowercase(sal) )
-
-domain_name()="$(domain_word()).$(domain_suffix())"
-
-company_email()= "$(user_name())@$(domain_name())"
-
-email()= executor(data["faker"]["internet"]["email_formats"])
-
-
-url()= executor(data["faker"]["internet"]["url_formats"])
-
-
-ipv4()= "$(rand( 172:255)).$(rand( 0:255)).$(rand(0:255)).$(rand(0:255))"
-
-ipv6()= (res = "";
-    for i = 1:8
-        res*= "$(string(rand( 4100:65535),base=16))"
-        if i < 8
-            res*= ":"
-        end
-    end;res)
-
-
-mac_address()=
-    (res = "";
-    for i = 1:6
-        res*= "$(Base.hex(unsigned(rand( 0x00:0xff)),2,false))"
-        i < 6 && (res*= ":")
-    end;res)
-
-uri_page()= executor(uri_pages)
-
-uri_path(deep="None")=
-    (if deep =="None"
+uri_path(;deep::Int=0)::String =
+    (if deep == 0
         deep = rand(1:3)
-    end; sa="";
-        for i =0:deep
-           sa*= uri_paths[rand( 1:length(uri_paths))]
-            i < deep && (sa*= "/")
-        end; sa )
+    end;
+    return join(map( x -> string(rand(uri_paths)) , 0:deep), "/") )
 
-uri_extension()= executor(uri_extensions)
-uri()= executor(data["faker"]["internet"]["uri_formats"])
-image_url()= executor(data["faker"]["internet"]["image_placeholder_services"])
+uri_extension()::String = rand(uri_extensions)
+uri()::String = executor(data["faker"]["internet"]["uri_formats"])
+image_url()::String = executor(data["faker"]["internet"]["image_placeholder_services"])